separate methods and tests into separate files for clarity
